Type matrix getters as double to get rid of unnecessary casts

+++ lib/src/filters/normal_map_filter.dart
@@ -216,12 +216,12 @@ class NormalMapFilterProgram extends RenderProgram {
     for (var i = 0, o = 0; i < vertexCount; i++, o += 4) {
       num x = vxList[o + 0];
       num y = vxList[o + 1];
-      vxData[vxIndex + 00] = posMatrix.tx + x * posMatrix.a + y * (posMatrix.c as double);
-      vxData[vxIndex + 01] = posMatrix.ty + x * posMatrix.b + y * (posMatrix.d as double);
-      vxData[vxIndex + 02] = texMatrix.tx + x * texMatrix.a + y * (texMatrix.c as double);
-      vxData[vxIndex + 03] = texMatrix.ty + x * texMatrix.b + y * (texMatrix.d as double);
-      vxData[vxIndex + 04] = mapMatrix.tx + x * mapMatrix.a + y * (mapMatrix.c as double);
-      vxData[vxIndex + 05] = mapMatrix.ty + x * mapMatrix.b + y * (mapMatrix.d as double);
+      vxData[vxIndex + 00] = posMatrix.tx + x * posMatrix.a + y * posMatrix.c;
+      vxData[vxIndex + 01] = posMatrix.ty + x * posMatrix.b + y * posMatrix.d;
+      vxData[vxIndex + 02] = texMatrix.tx + x * texMatrix.a + y * texMatrix.c;
+      vxData[vxIndex + 03] = texMatrix.ty + x * texMatrix.b + y * texMatrix.d;
+      vxData[vxIndex + 04] = mapMatrix.tx + x * mapMatrix.a + y * mapMatrix.c;
+      vxData[vxIndex + 05] = mapMatrix.ty + x * mapMatrix.b + y * mapMatrix.d;
       vxData[vxIndex + 06] = ambientR;
       vxData[vxIndex + 07] = ambientG;
       vxData[vxIndex + 08] = ambientB;
@@ -230,8 +230,8 @@ class NormalMapFilterProgram extends RenderProgram {
       vxData[vxIndex + 11] = lightG;
       vxData[vxIndex + 12] = lightB;
       vxData[vxIndex + 13] = lightA;
-      vxData[vxIndex + 14] = lightX as double;
-      vxData[vxIndex + 15] = lightY as double;
+      vxData[vxIndex + 14] = lightX;
+      vxData[vxIndex + 15] = lightY;
       vxData[vxIndex + 16] = lightZ;
       vxData[vxIndex + 17] = lightRadius;
       vxData[vxIndex + 18] = alpha;
